# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crm/tasks:
    get:
      operationId: crmGetTasks
      tags: [crm-tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmTask"
    post:
      operationId: crmCreateTask
      tags: [crm-tasks]
      summary: Create a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmTask"
  /crm/tasks/{id}:
    patch:
      operationId: crmUpdateTask
      tags: [crm-tasks]
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmTaskInput"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmTask"
    delete:
      operationId: crmDeleteTask
      tags: [crm-tasks]
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"

  # ── CRM Users (admin only) ───────────────────────────
  /crm/users:
    get:
      operationId: crmGetUsers
      tags: [crm-users]
      summary: List CRM users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmUser"
    post:
      operationId: crmCreateUser
      tags: [crm-users]
      summary: Create a CRM user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmUserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmUser"
  /crm/users/{id}:
    patch:
      operationId: crmUpdateUser
      tags: [crm-users]
      summary: Update a CRM user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmUserInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmUser"
    delete:
      operationId: crmDeleteUser
      tags: [crm-users]
      summary: Delete a CRM user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"

  # ── CRM Submission Links ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    CrmUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, sales, va]
        status:
          type: string
          enum: [active, inactive]
        campaignId:
          type: integer
          nullable: true
        campaignName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, status, createdAt]
    CrmUserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, sales, va]
        status:
          type: string
          enum: [active, inactive]
        campaignId:
          type: integer
          nullable: true
      required: [name, email, role]
...
    CrmTask:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
          nullable: true
        leadAddress:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, completed]
        createdAt:
          type: string
      required: [id, title, status, createdAt]
    CrmTaskInput:
      type: object
      properties:
        leadId:
          type: integer
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, completed]
      required: [title]
...